bot/requirements.txt
aiogram==3.2.0
aiohttp==3.8.6
python-decouple==3.8

